# DataViews: add actions to list layout