--- a/packages/dataviews/src/style.scss
+++ b/packages/dataviews/src/style.scss
@@ -406,8 +406,14 @@
 			position: relative;
 			border-radius: $grid-unit-05;
 
-			> * {
-				width: 100%;
+			&::after {
+				position: absolute;
+				content: "";
+				top: 100%;
+				left: $grid-unit-30;
+				right: $grid-unit-30;
+				background: $gray-100;
+				height: 1px;
 			}
 		}