--- a/packages/dataviews/src/style.scss
+++ b/packages/dataviews/src/style.scss
@@ -412,6 +412,10 @@
 			}
 		}
 
+		.dataviews-item-actions .components-button:not(.dataviews-all-actions-button) {
+			opacity: 0;
+		}
+
 		&:not(.is-selected) {
 			.dataviews-view-list__primary-field {
 				color: $gray-900;
@@ -425,8 +429,12 @@
 				.dataviews-view-list__fields {
 					color: var(--wp-admin-theme-color);
 				}
+				.dataviews-item-actions .components-button:not(.dataviews-all-actions-button) {
+					opacity: 1;
+				}
 			}
 		}
+
 	}
 
 	li.is-selected,