--- a/packages/dataviews/src/style.scss
+++ b/packages/dataviews/src/style.scss
@@ -416,6 +416,13 @@
 			opacity: 0;
 		}
 
+		&.is-hovered,
+		&:hover {
+			.dataviews-item-actions .components-button:not(.dataviews-all-actions-button) {
+				opacity: 1;
+			}
+		}
+
 		&:not(.is-selected) {
 			.dataviews-view-list__primary-field {
 				color: $gray-900;
@@ -428,9 +435,6 @@
 				.dataviews-view-list__primary-field,
 				.dataviews-view-list__fields {
 					color: var(--wp-admin-theme-color);
-				}
-				.dataviews-item-actions .components-button:not(.dataviews-all-actions-button) {
-					opacity: 1;
 				}
 			}
 		}