--- a/packages/dataviews/src/style.scss
+++ b/packages/dataviews/src/style.scss
@@ -528,6 +528,7 @@
 	}
 
 	.dataviews-item-actions {
+		padding-top: $grid-unit-20;
 		padding-right: $grid-unit-40;
 	}