--- a/packages/dataviews/src/style.scss
+++ b/packages/dataviews/src/style.scss
@@ -514,6 +514,10 @@
 		}
 	}
 
+	.dataviews-item-actions {
+		padding-right: $grid-unit-40;
+	}
+
 	& + .dataviews-pagination {
 		justify-content: space-between;
 	}