--- a/packages/dataviews/src/style.scss
+++ b/packages/dataviews/src/style.scss
@@ -452,7 +452,7 @@
 	}
 
 	.dataviews-view-list__item {
-		padding: $grid-unit-20 $grid-unit-40;
+		padding: $grid-unit-20 0 $grid-unit-20 $grid-unit-40;
 		width: 100%;
 		scroll-margin: $grid-unit-10 0;