@oandregal 2024-04-17
> Keyboard interaction: the actions button should be reachable via the left arrow key, not by tabbing.

https://github.com/WordPress/gutenberg/pull/59637 introduced arrow-key navigation for the list items. I don't remember the details, but my expectation was that there was a single tab stop for the list items (the whole item) and the actions button (called details button at the time) was reachable via arrow keys (right/left).

Code changed a bit since and this PR re-introduces the actions button, though it's not working as I expected (note how the tab navigation actually goes into the action button and right-left arrow keys don't work):

https://github.com/WordPress/gutenberg/assets/583546/d81c6f4a-b0ed-48e8-993d-0d050335ce0a