@oandregal 2024-04-30
By adding a second `<CompositeItem>` within the same `<CompositeRow>` this PR is trying to add two-dimensional arrow key navigation:

- top/bottom arrow navigates up/down through the list of items
- left/right arrow navigates within a single item: item <=> actions button

According to the AriaKit's docs for Composite, this is the model we should be using:

```