@DaniGuardiola 2024-05-06
Hi there @oandregal 👋 😄 

I want to help. Two questions:

1. Did you find a solution or workaround?
2. Did you report this in the Ariakit repo? If so, can you link it?

cc @mirka as I believe you were looking into this.

I could take a look and try to fix this in Ariakit's side.