@oandregal 2024-05-06
Hi @DaniGuardiola thanks for offering help :)

Just this morning was tinkering with the example Lena shared [in slack](https://wordpress.slack.com/archives/C02QB2JS7/p1714554441286959?thread_ts=1714479772.233349&cid=C02QB2JS7) using only Ariaki components, and modified it so it has [multiple rows](https://stackblitz.com/edit/vitejs-vite-neycie?file=src%2FApp.tsx&terminal=dev). It demonstrates how Menus don't work well with composite (or there's some config that I'm missing).

I've looked into Ariakit's Github, and didn't find any issue for this. The only issue related to composite is https://github.com/ariakit/ariakit/issues/3170 created by Andrew, but it's a bit different one. I'm creating an issue now.