@oandregal 2024-05-08
As per the conversation at https://github.com/ariakit/ariakit/issues/3768 and this example by Dani https://stackblitz.com/edit/vitejs-vite-p27qhy?file=src%2FApp.tsx&terminal=dev there seems to be a way to make two dimensional arrow key navigation.

Specifically, this bit is the key:

```jsx
          <Ariakit.CompositeRow>
            <Ariakit.CompositeItem render={<button>Button</button>} />
            <Ariakit.MenuProvider>
              {/* here, CompositeItem has precedence and pressing arrow down
              goes downwards in the 2d composite as expected */}
              <Ariakit.CompositeItem
                store={store}
                render={<Ariakit.MenuButton>Menu</Ariakit.MenuButton>}
              />
              <Ariakit.Menu>
                <Ariakit.MenuItem>Hello</Ariakit.MenuItem>
              </Ariakit.Menu>
            </Ariakit.MenuProvider>
```

I'm looking into how we can achieve the same with the existing `@wordpress/components` we have.