@oandregal 2024-05-08
Using the `Composite*` items from `@wordpress/components` as of e97be2b101cb28f887e0b54e5286d00ac93fd392 still works as expected. The last bit is substituting the following by `@wordpress/components`:

```jsx
<Ariakit.MenuProvider>
    <CompositeItem
        store={ store }
        render={
            <Ariakit.MenuButton>Menu</Ariakit.MenuButton>
        }
    />
    <Ariakit.Menu>
        <Ariakit.MenuItem>Hello</Ariakit.MenuItem>
    </Ariakit.Menu>
</Ariakit.MenuProvider>
```