@oandregal 2024-05-08
As of https://github.com/WordPress/gutenberg/pull/60805/commits/f10e368fd9de4f1e8b7c277332a354773b5a1b7c I've got it working with `@wordpress/components`:

https://github.com/WordPress/gutenberg/assets/583546/a48b37fb-707e-4e8f-aacc-bc1cc1d6a31a

It's not complete, I need to add support for primary actions, verify everything is working properly, etc. Also see if there's ways to modify `ItemActions` so I can reuse here — as soon as we start improving the other layouts to add arrow key navigation I presume we'll run into the same issues with this component.