@tyxla 2024-05-08
> > I'm not sure why we want to use wordpress components here. Are the styles the exact same? These components are little more than themed and slightly more abstracted Ariakit components, not necessarily meant to be low-level. I think using Ariakit components here directly is fine unless you need to reuse a significant mount of styles from the wordpress component in question.
> 
> As far as my understanding goes, using Ariakit directly is not allowed — there's even lint rules that prevent you from committing if you do so. I'm not sure why this is. Anyway, I've got this PR working with `@wordpress/components` :)

This is correct and very intentional. The `@wordpress/components` package uses Ariakit under the hood, but we aim to keep this an implementation detail and not something that the end user has to care about. For various reasons, we want to add custom layers on top of Ariakit, and using Ariakit components directly might prevent the consumers from leveraging the benefits of those layers. Keeping those internals private is also helpful in case we decide to start using a different underlying library for component X.