@oandregal 2024-05-09
Interaction-wise, this is working now with every kind of action (modal & non modal):

https://github.com/WordPress/gutenberg/assets/583546/b86966b5-6f58-4a64-8819-0ff867f077e9

I still have to test all action flows and verify that everything works properly. I'm also looking for opportunities to improve the existing code now that it's working, so things can change a bit code-wise.

@youknowriad @jameskoster While I do that, I'd welcome feedback on the following:

> What primary actions should we display in Pages & Templates in the list view? The same as in any other view or limit it only to Edit?

| Page | Primary actions for Table layout | Primary actions for Grid layout | Primary actions for List layout |
| --- | --- | --- | --- |
| Templates | Edit | - | Edit |
| Pages | Edit, View, Trash | - | Edit |

The original issue suggested only Edit https://github.com/WordPress/gutenberg/issues/59659 and that sounds good to me, but I wanted to double check. I plan to implement this in consumer-land, not as a new dataviews API: this is, the Pages component will change the primary actions depending on layout.