@oandregal 2024-05-15
> Even though the "navigate" metric in the site editor is a bit unstable in our tests, I do think this PR had a small impact on the numbers there. I wonder if you have ideas.

Yeah, I agree. I've tried before/after locally and I see a consistent +15ms after this PR. This is what I've looked at (the same as the navigate performance test):

- set up `SCRIPT_DEBUG` & `WP_DEBUG` to false in the env
- visit the templates page & switched to list layout
- start performance trace and click on "Single posts" item, so the preview is updated
- inspect the duration of the click event 

Codehealth after this commit reports a higher mean number as well:

<img width="2102" alt="Captura de ecrã 2024-05-15, às 12 34 59" src="https://github.com/WordPress/gutenberg/assets/583546/d65b476d-df45-4f3e-acd3-590485085958">

I've looked at browser's performance traces and they are very similar ([before.json.zip](https://github.com/WordPress/gutenberg/files/15320075/before.json.zip) & [after.json.zip](https://github.com/WordPress/gutenberg/files/15320080/after.json.zip), for anyone to import into devtools and inspect):

| Before | After |
| --- | --- |
| <img width="560" alt="Captura de ecrã 2024-05-15, às 12 33 19" src="https://github.com/WordPress/gutenberg/assets/583546/a57fbf55-5171-40c1-a5c6-2f8481d15035"> | <img width="560" alt="Captura de ecrã 2024-05-15, às 12 33 05" src="https://github.com/WordPress/gutenberg/assets/583546/85887fb7-40be-4a7f-bc8f-d5828e0dd4a1"> |

The major thing I've noticed is the [Role component](https://github.com/ariakit/ariakit/blob/cdf65c27bb928412ae0dee0cccf9b3fdb229515c/packages/ariakit-react-core/src/utils/system.tsx#L13) coming from Ariakit's reac-core package (imported from `@wordpress/components`). While it exists in both traces, it takes +8ms more in the after trace. What's intriguing is that it's not called more times, it just takes more time to do the same number of calls:

| Before | After |
| --- | --- |
| <img width="1125" alt="Captura de ecrã 2024-05-15, às 13 07 14" src="https://github.com/WordPress/gutenberg/assets/583546/e7ca6c0d-eaf1-4b08-831c-f41d2762c385"> | <img width="1125" alt="Captura de ecrã 2024-05-15, às 13 06 57" src="https://github.com/WordPress/gutenberg/assets/583546/36f99c85-e5b0-4885-b3ba-dbdb1f47f485"> |

So I guess it's fair to say that the performance hit comes from the internal components/ariakit code, which is in line with what this PR does:

- adds dropdown from `@wordpress/components` that also uses ariakit under the hood
- adds more composite items (primary action + all actions buttons)
- uses the composite store to do some things

Further debugging would require looking more deeply at how `@wordpress/components` and Ariakit works, to understand how it can be improved. Hope this serves as a starting point.